working API V1

+++ portal_roa_accums.py
@@ -6,6 +6,7 @@
 #               for accPeriod 24h @ 0800, save .nc file to nowcast dir and also 
 #               accumulate the past 30 days.nc files to create the API
 #         SRA changed average to include the endpoint (for trepezioidal averaging)
+#          SRA Added API calculations and plotting functionality
 #"""
 import rasterio
 import xarray as xr
@@ -15,6 +16,10 @@ import netCDF4 as nc
 import datetime,time
 import os,glob,sys,argparse
 from itertools import chain
+import matplotlib.pyplot as plt
+import matplotlib as mpl
+from matplotlib.colors import from_levels_and_colors
+from mpl_toolkits.basemap import Basemap
 
 nrows = 2962
 ncols=  2777
@@ -22,13 +27,14 @@ delta_y = 0.02701789935
 delta_x = 0.02701729106
 
 dataDir='/mnt/prj/swift/rain_over_africa'
-#tmpDir='/home/stewells/AfricaNowcasting/tmp/'
-tmpDir='/mnt/users/hymod/seodey/NFLICS/RoA_files/tmp/'
-#geotiffDir = '/mnt/HYDROLOGY_stewells/geotiff/ssa_africarain_precip_accum/'
-geotiffDir = '/mnt/users/hymod/seodey/NFLICS/RoA_files/ssa_africarain_precip_accum/'
-#backupDir = '/mnt/data/hmf/projects/LAWIS/WestAfrica_portal/SANS_transfer/data'
-backupDir = '/mnt/users/hymod/seodey/NFLICS/RoA_files/backup'
-nflicsDir='/mnt/users/hymod/seodey/NFLICS/nflics_nowcasts/'   #for saving the .nc files for calculating the API
+tmpDir='/home/stewells/AfricaNowcasting/tmp/'
+#tmpDir='/mnt/users/hymod/seodey/NFLICS/RoA_files/tmp/'
+geotiffDir = '/mnt/HYDROLOGY_stewells/geotiff/ssa_africarain_precip_accum/'
+#geotiffDir = '/mnt/users/hymod/seodey/NFLICS/RoA_files/ssa_africarain_precip_accum/'
+backupDir = '/mnt/data/hmf/projects/LAWIS/WestAfrica_portal/SANS_transfer/data'
+#backupDir = '/mnt/users/hymod/seodey/NFLICS/RoA_files/backup'
+#nflicsDir='/mnt/users/hymod/seodey/NFLICS/nflics_nowcasts/'   #for saving the .nc files for calculating the API
+nflicsDir='/mnt/prj/nflics/nflics_nowcasts/' 
 
 #geotiffDir = '/home/stewells/AfricaNowcasting/satTest/'
 testDate = '202406201300'
@@ -38,7 +44,8 @@ outputNChr=8
 outputNCacc=24
 
 toSdir = False
-
+saveNC=True   #save outputNCacc at outpuNChr to netCDF file
+doAPI=True      #caluclate and save the 30 day API
 
 def make_geoTiff(data,rasFile,doReproj = True,origEPSG='4326',newEPSG='3857',reprojFile='test.tif',trim=False):
     nbands = len(data)
@@ -88,6 +95,28 @@ def generate_dates(start,end,interval):
         current+= delta
     return dateList
 
+def plot_roa_acc(toplt,outfile,cbarlab,plot_lims,roa_grid,use_title):
+    rcmap,rnorm = from_levels_and_colors([-10000,0,2,10,50,100,150,200],
+            ['#666363','#EBE5E4','#129FFE','#00FF40','#FFFF00','#FF8000','#FF0000','#FF0080'],extend="max") 
+    fig, ax1 = plt.subplots()
+    m = Basemap(projection='merc',ax=ax1,lat_0=0.,lon_0=0., resolution="h",
+            llcrnrlon=plot_lims[1],llcrnrlat=plot_lims[0],
+            urcrnrlon=plot_lims[3],urcrnrlat=plot_lims[2])        
+    
+    X, Y = m(roa_grid[0],roa_grid[1])
+    pc = m.pcolormesh(X,Y,toplt,cmap=rcmap,norm=rnorm)
+    m.drawcoastlines(linewidth=0.5)
+    m.drawcountries(linewidth=0.5)
+    cbr=plt.colorbar(pc,pad=0.08,shrink=0.49,aspect=10,extend="max")        
+    cbr.ax.set_title(cbarlab,fontsize=10)
+    m.drawparallels(np.arange(-50,50,10),labels=[1,0,0,1])
+    m.drawmeridians(np.arange(-50,50,10),labels=[1,0,0,1])
+    plt.title(use_title,pad=1,fontsize=10)   
+    #cr=m.contour(Xmid,Ymid,blobs,colors='cyan',linewidths=2,linestyles="solid",levels=range(14,220,15))
+    fig.set_size_inches(6,6)  
+    plt.tight_layout(pad=1)
+    plt.savefig(outfile)
+    
 def getAccs(tnow,accPeriods,dataDir,tmpDir,geotiffDir):
     print("Generating accumulations for "+str(tnow))
     tnowStr = tnow
@@ -134,7 +163,7 @@ def getAccs(tnow,accPeriods,dataDir,tmpDir,geotiffDir):
             make_geoTiff([accArr_i],rasPath,reprojFile=rasPath_3857,trim=True)
             os.system('rm '+rasPath)
 
-            if tnow.hour==outputNChr and tnow.minute==0 and acchr ==outputNCacc:
+            if saveNC and tnow.hour==outputNChr and tnow.minute==0 and acchr ==outputNCacc:
                 print("saving .nc output for hour-of-day",outputNChr,"acc", acchr, "h") 
                 ds=xr.Dataset() #create dataset to save to netcdf for future use
                 ds['posterior_mean_24h']=xr.DataArray(accArr_i, coords={'latitude': dfile.coords["latitude"].data , 'longitude': dfile.coords["longitude"].data},dims=['latitude', 'longitude']) 
@@ -153,45 +182,56 @@ def getAccs(tnow,accPeriods,dataDir,tmpDir,geotiffDir):
                 if not os.path.exists(outpathNC):
                     os.makedirs(outpathNC)            
                 ds.to_netcdf(path=outfileNC,mode='w', encoding=enc, format='NETCDF4') 
- 
+                #print("plotting")
+                roa_grid=np.meshgrid(dfile.coords["longitude"],dfile.coords["latitude"])
+                #plot_roa_acc(accArr_i,outfileNC.split(".")[0]+".png","Precip. [mm]",[-40,-20,33,50],roa_grid,tnowStr)
+                
                 #second calculate the API 
-                API_w=np.power(0.9,np.arange(0,30,1))
-                API=[accArr_i]
-                #add the weighted sum of previous 30 DAILY accumulations
-                for iAPI,dAPI in enumerate([tnow - n*datetime.timedelta(hours=24)  for n in np.arange(1,30)]):
-                    print(iAPI, dAPI)
-                    outpathNCi=os.path.join(nflicsDir,str(dAPI.year), str(dAPI.month).zfill(2),str(dAPI.day).zfill(2))
-                    outfileNCi=os.path.join(outpathNCi,"rainoverAfrica_SSA_"+dAPI.strftime('%Y%m%d%H%M')+"_acc"+str(acchr)+"h.nc")
-                    dsi=xr.open_dataset(outfileNCi)
-                    API.append(dsi["posterior_mean_24h"]*API_w[iAPI+1])
-                API=np.array(API)
-                APIsum=np.sum(API, axis=0)
-                
-                #save as .tiff and .nc for checking
-                print("saving .nc output for hour-of-day",outputNChr,"acc", acchr, "h") 
-                ds=xr.Dataset() #create dataset to save to netcdf for future use
-                ds['API']=xr.DataArray(APIsum, coords={'latitude': dfile.coords["latitude"].data , 'longitude': dfile.coords["longitude"].data},dims=['latitude', 'longitude']) 
-                ds.attrs["calulation"]="0.9^n-1*Pn for n =1:30 is number of days before today"
-                ds.attrs["units"]="mm"
-                
-                comp = dict(zlib=True, complevel=5)
-                enc = {var: comp for var in ds.data_vars}
-                outpathNC=os.path.join(nflicsDir,str(tnow.year), str(tnow.month).zfill(2),str(tnow.day).zfill(2))
-                outfileNC=os.path.join(outpathNC,"rainoverAfrica_SSA_"+tnowStr+"_API30_"+str(tnow.hour).zfill(2)+"00.nc")
-                print("saving .nc output to", outfileNC)
-                if not os.path.exists(outpathNC):
-                    os.makedirs(outpathNC)            
-                ds.to_netcdf(path=outfileNC,mode='w', encoding=enc, format='NETCDF4') 
-                
-                #calculate and save .tiff
-                rasPath = os.path.join(tmpDir,"ROA_precip_API_"+tnowStr+"_SSA.tif")
-                if toSdir:
-                    rasPath_3857 = os.path.join(backupDir,"ROA_precip_API_"+tnowStr+"_3857.tif")
-                else:
-                    rasPath_3857 = os.path.join(geotiffDir,tnow.strftime('%Y%m%d'),"ROA_precip_API_"+tnowStr+"_3857.tif")
-                
-                make_geoTiff([APIsum],rasPath,reprojFile=rasPath_3857,trim=True)
-                os.system('rm '+rasPath)
+                if doAPI:
+                    API_w=np.power(0.9,np.arange(0,30,1))
+                    API=[accArr_i]
+                    #add the weighted sum of previous 30 DAILY accumulations
+                    for iAPI,dAPI in enumerate([tnow - n*datetime.timedelta(hours=24)  for n in np.arange(1,30)]):
+                        #print(iAPI, dAPI)
+                        outpathNCi=os.path.join(nflicsDir,str(dAPI.year), str(dAPI.month).zfill(2),str(dAPI.day).zfill(2))
+                        outfileNCi=os.path.join(outpathNCi,"rainoverAfrica_SSA_"+dAPI.strftime('%Y%m%d%H%M')+"_acc"+str(acchr)+"h.nc")
+                        try:
+                            dsi=xr.open_dataset(outfileNCi)
+                            API.append(dsi["posterior_mean_24h"]*API_w[iAPI+1])
+                        except:
+                            print("Missing file "+outfileNCi)
+                    API=np.array(API)
+                    if len(API)>1: 
+                        API=np.array(API)
+                        APIsum=np.sum(API, axis=0)
+                    else:
+                        API=np.array(API)
+                    #save as .tiff and .nc for checking
+                    print("saving .nc output for hour-of-day",outputNChr,"acc", acchr, "h") 
+                    ds=xr.Dataset() #create dataset to save to netcdf for future use
+                    ds['API']=xr.DataArray(APIsum, coords={'latitude': dfile.coords["latitude"].data , 'longitude': dfile.coords["longitude"].data},dims=['latitude', 'longitude']) 
+                    ds.attrs["calulation"]="0.9^n-1*Pn for n =1:30 is number of days before today"
+                    ds.attrs["units"]="mm"
+                    
+                    comp = dict(zlib=True, complevel=5)
+                    enc = {var: comp for var in ds.data_vars}
+                    outpathNC=os.path.join(nflicsDir,str(tnow.year), str(tnow.month).zfill(2),str(tnow.day).zfill(2))
+                    outfileNC=os.path.join(outpathNC,"rainoverAfrica_SSA_"+tnowStr+"_API30_"+str(tnow.hour).zfill(2)+"00.nc")
+                    print("saving .nc output to", outfileNC)
+                    if not os.path.exists(outpathNC):
+                        os.makedirs(outpathNC)            
+                    ds.to_netcdf(path=outfileNC,mode='w', encoding=enc, format='NETCDF4')                 
+                    plot_roa_acc(APIsum,outfileNC.split(".")[0]+".png","API30 [mm]",[-40,-20,33,50],roa_grid,tnowStr)
+
+                    #calculate and save .tiff
+                    rasPath = os.path.join(tmpDir,"ROA_precip_API_"+tnowStr+"_SSA.tif")
+                    if toSdir:
+                        rasPath_3857 = os.path.join(backupDir,"ROA_precip_API_"+tnowStr+"_3857.tif")
+                    else:
+                        rasPath_3857 = os.path.join(geotiffDir,tnow.strftime('%Y%m%d'),"ROA_precip_API_"+tnowStr+"_3857.tif")
+                    
+                    make_geoTiff([APIsum],rasPath,reprojFile=rasPath_3857,trim=True)
+                    os.system('rm '+rasPath)
                 
 #   print(ifile)
 
@@ -274,7 +314,7 @@ if __name__ == '__main__':
             print("start date rounded to nearest interval matching raw data")
             endDate = round_edate
         # get list of dates
-        dateList = generate_dates(startDate,endDate,1440)   #SRA EDITED FOR DAILY PROCESSING - CHANGE BACK!!!!!
+        dateList = generate_dates(startDate,endDate,15)   #SRA EDITED FOR DAILY PROCESSING - CHANGE BACK!!!!!
         new_files = [x.strftime('%Y%m%d%H%M') for x in dateList if reprocess or not os.path.exists(os.path.join(geotiffDir,x.strftime("%Y%m%d"),x.strftime('rainoverAfrica_SSA_%Y%m%d%H%M_acc1h_3857.tif')))]
 
     print(new_files)
